% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarEffects.R
\docType{data}
\name{calendarEffects}
\alias{calendarEffects}
\title{Brazilian calendar effects}
\format{A \code{list} with 2 elements: 
\itemize{
 \item \code{x} is a \code{mts} with 4 series: td (trading days), easter, carnival an leapYear;
 \item \code{desc} is a \code{data.frame} with specifications of the series in \code{x}.
 }}
\source{
The calendar data was collected in IPEA Data \url{www.ipeadata.gov.br}.
}
\usage{
calendarEffects
}
\description{
Brazilian calendar variables to use in seasonal adjustment.
}
