% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagX13.R
\name{diagX13}
\alias{diagX13}
\title{SACE diagnostic tool}
\usage{
diagX13(x, file = "")
}
\arguments{
\item{x}{output from seasX13 function}

\item{file}{a character string naming the file}
}
\description{
The function uses some tests developed specifically for the SACE series to diagnose the seasonal adjustment.
}
