% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listModels.R
\docType{data}
\name{listModels}
\alias{listModels}
\title{List of models to correct time series seasonal adjustment}
\format{A \code{data.frame} with 48 possibles models.}
\usage{
listModels
}
\description{
List of possible models to time series seasonal adjustment when using the argument \code{autoCorrection} in seasX13 function.
}
