% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMECX13.R
\docType{package}
\name{NMECX13}
\alias{NMECX13}
\alias{NMECX13-package}
\title{NMECX13: a package to brazilian time series seasonal adjustment.}
\description{
The NMECX13 package uses X-13ARIMA-SEATS program by US Census Bureau to perform seasonal adjustment in multiple time series simultaneously.
}
\note{
We use the `seasonal` R package developed by Christopher Sax to perform the seasonal adjustmente routine.
}
\author{
Daiane Marcolino de Mattos \email{daiane.mattos@fgv.br},
        Bianca Gonçalves \email{biagbp@gmail.com},
        Pedro Guilherme Costa Ferreira \email{pedro.guilherme@fgv.br}
}
