% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readX13.R
\name{readX13}
\alias{readX13}
\title{Read data file for seasonal adjustment}
\usage{
readX13(path = "", sheetNumber = 1)
}
\arguments{
\item{path}{path to the csv/xlsx file}

\item{sheetNumber}{sheet number of xlsx file}
}
\value{
A \code{list} containing the following elements:
\item{xts}{time series in the path file.}
\item{xtsNA}{a object identifying the missing observations in each series.}
\item{deniedNames}{a vector naming the time series that will not be seasonally adjusted (less than three years of observation).}
\item{acceptedNames}{a vector naming time series that can be seasonally adjusted.}
\item{path}{path to the csv/xlsx file}
}
\description{
Read a CSV or XLSX file. The file must have 2 or more columns. The first one must contain the sequential date of the time series. Missing values are supported.
}
