% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionX13.R
\name{regressionX13}
\alias{regressionX13}
\title{SACE regression tool}
\usage{
regressionX13(x, series, espec = NULL, file = NULL)
}
\arguments{
\item{x}{output from readX13 function}

\item{series}{a vector string naming the series you want to obtain the seasonal factors}

\item{espec}{a vector string that indicates the series transformation: log or none}

\item{file}{a character string naming the file}
}
\description{
The function estimates a regression to obtain the seasonal factors.
}
