\name{saveX13}
\alias{saveX13}
\title{Save seasonal adjustment results in multiple files}
\usage{
saveX13(output, file = "")
}
\arguments{
\item{output}{output from seasX13 function}
\item{file}{a character string naming a file}
}

\description{
Save CSV files for seasonally adjusted data, seasonal factors, calendar factors, total factors (seasonal & calendar factors) and model specification for each series.
}

\examples{

}
