% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceSurvey.R
\docType{data}
\name{serviceSurvey}
\alias{serviceSurvey}
\title{Confidence index from Service Survey (FGV IBRE)}
\format{A \code{data.frame} with 4 columns: the first one is in \code{Date} format, the others are in \code{numeric}.}
\source{
The Dataset was collected from Service Survey (Getulio Vargas Foundation - Brazilian Institute of Economics).
}
\usage{
serviceSurvey
}
\description{
file example accepted in the seasonal adjustment routine.
}
