% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveX13.R
\name{saveX13}
\alias{saveX13}
\title{Save seasonal adjustment results in multiple files}
\usage{
saveX13(output, file = "")
}
\arguments{
\item{output}{output from seasX13 function}

\item{file}{a character string naming the file}
}
\description{
Export CSV files for seasonally adjusted data, seasonal factors, calendar factors, total factors (seasonal & calendar factors) and model specification for each series.
}
\examples{
\dontrun{
# read data example, auto seasonal adjustment, export results
data(serviceSurvey)
data <- readX13(serviceSurvey)
auto <- seasX13(data)

saveX13(auto, "auto")
}
}
